+++ tests/test_database.py
@@ -1,11 +1,23 @@
 from contextlib import asynccontextmanager, contextmanager
 from unittest.mock import MagicMock
 
+import cloudpickle
 import pytest
 from prefect import flow
+from sqlalchemy.engine import Connection, Engine
+from sqlalchemy.ext.asyncio import AsyncConnection, AsyncEngine
 
-from prefect_sqlalchemy.credentials import AsyncDriver, DatabaseCredentials, SyncDriver
-from prefect_sqlalchemy.database import sqlalchemy_execute, sqlalchemy_query
+from prefect_sqlalchemy.credentials import (
+    AsyncDriver,
+    ConnectionComponents,
+    DatabaseCredentials,
+    SyncDriver,
+)
+from prefect_sqlalchemy.database import (
+    SqlAlchemyConnector,
+    sqlalchemy_execute,
+    sqlalchemy_query,
+)
 
 
 class SQLAlchemyAsyncEngineMock:
@@ -230,3 +242,235 @@ async def test_sqlalchemy_execute_sqlite_async(tmp_path):
     actual_names, actual_addresses = zip(*rows)
     assert expected_names == set(actual_names)
     assert expected_address == actual_addresses[0]
+
+
+class TestSqlAlchemyConnector:
+    async def test_connector_init(self):
+        credentials_components = SqlAlchemyConnector(
+            connection_info=ConnectionComponents(
+                driver=SyncDriver.POSTGRESQL_PSYCOPG2,
+                username="myusername",
+                password="mypass",
+                database="my.db",
+                host="localhost",
+                port=1234,
+            ),
+        )
+
+        connection_url = "postgresql+psycopg2://myusername:mypass@localhost:1234/my.db"
+        credentials_url = SqlAlchemyConnector(connection_info=connection_url)
+        assert credentials_components._rendered_url == credentials_url._rendered_url
+
+    @pytest.fixture(params=[SyncDriver.SQLITE_PYSQLITE, AsyncDriver.SQLITE_AIOSQLITE])
+    async def connector_with_data(self, tmp_path, request):
+        credentials = SqlAlchemyConnector(
+            connection_info=ConnectionComponents(
+                driver=request.param,
+                database=str(tmp_path / "test.db"),
+            ),
+            fetch_size=2,
+        )
+        await credentials.execute(
+            "CREATE TABLE IF NOT EXISTS customers (name varchar, address varchar);"
+        )
+        await credentials.execute(
+            "INSERT INTO customers (name, address) VALUES (:name, :address);",
+            parameters={"name": "Marvin", "address": "Highway 42"},
+        )
+        await credentials.execute_many(
+            "INSERT INTO customers (name, address) VALUES (:name, :address);",
+            seq_of_parameters=[
+                {"name": "Ford", "address": "Highway 42"},
+                {"name": "Unknown", "address": "Space"},
+                {"name": "Me", "address": "Myway 88"},
+            ],
+        )
+        yield credentials
+
+    @pytest.fixture(params=[True, False])
+    async def managed_connector_with_data(self, connector_with_data, request):
+        if request.param:
+            if connector_with_data._driver_is_async:
+                async with connector_with_data:
+                    yield connector_with_data
+            else:
+                with connector_with_data:
+                    yield connector_with_data
+        else:
+            yield connector_with_data
+            if connector_with_data._driver_is_async:
+                await connector_with_data.aclose()
+            else:
+                connector_with_data.close()
+        assert connector_with_data._unique_results == {}
+        assert connector_with_data._engine is None
+
+    @pytest.mark.parametrize("begin", [True, False])
+    def test_get_connection(self, begin, managed_connector_with_data):
+        connection = managed_connector_with_data.get_connection(begin=begin)
+        if begin:
+            engine_type = (
+                AsyncEngine if managed_connector_with_data._driver_is_async else Engine
+            )
+            assert isinstance(connection, engine_type._trans_ctx)
+        else:
+            engine_type = (
+                AsyncConnection
+                if managed_connector_with_data._driver_is_async
+                else Connection
+            )
+            assert isinstance(connection, engine_type)
+
+    @pytest.mark.parametrize("begin", [True, False])
+    def test_get_client(self, begin, managed_connector_with_data):
+        connection = managed_connector_with_data.get_client(
+            client_type="connection", begin=begin
+        )
+        if begin:
+            engine_type = (
+                AsyncEngine if managed_connector_with_data._driver_is_async else Engine
+            )
+            assert isinstance(connection, engine_type._trans_ctx)
+        else:
+            engine_type = (
+                AsyncConnection
+                if managed_connector_with_data._driver_is_async
+                else Connection
+            )
+            assert isinstance(connection, engine_type)
+
+    async def test_reset_connections_sync_async_error(
+        self, managed_connector_with_data
+    ):
+        with pytest.raises(RuntimeError, match="synchronous connections"):
+            if managed_connector_with_data._driver_is_async:
+                managed_connector_with_data.reset_connections()
+            else:
+                await managed_connector_with_data.reset_async_connections()
+
+    async def test_fetch_one(self, managed_connector_with_data):
+        results = await managed_connector_with_data.fetch_one("SELECT * FROM customers")
+        assert results == ("Marvin", "Highway 42")
+        results = await managed_connector_with_data.fetch_one("SELECT * FROM customers")
+        assert results == ("Ford", "Highway 42")
+
+        # test with parameters
+        results = await managed_connector_with_data.fetch_one(
+            "SELECT * FROM customers WHERE address = :address",
+            parameters={"address": "Myway 88"},
+        )
+        assert results == ("Me", "Myway 88")
+        assert len(managed_connector_with_data._unique_results) == 2
+
+        # now reset so fetch starts at the first value again
+        if managed_connector_with_data._driver_is_async:
+            await managed_connector_with_data.reset_async_connections()
+        else:
+            managed_connector_with_data.reset_connections()
+        assert len(managed_connector_with_data._unique_results) == 0
+
+        # ensure it's really reset
+        results = await managed_connector_with_data.fetch_one("SELECT * FROM customers")
+        assert results == ("Marvin", "Highway 42")
+        assert len(managed_connector_with_data._unique_results) == 1
+
+    @pytest.mark.parametrize("size", [None, 1, 2])
+    async def test_fetch_many(self, managed_connector_with_data, size):
+        results = await managed_connector_with_data.fetch_many(
+            "SELECT * FROM customers", size=size
+        )
+        expected = [("Marvin", "Highway 42"), ("Ford", "Highway 42")][
+            : (size or managed_connector_with_data.fetch_size)
+        ]
+        assert results == expected
+
+        # test with parameters
+        results = await managed_connector_with_data.fetch_many(
+            "SELECT * FROM customers WHERE address = :address",
+            parameters={"address": "Myway 88"},
+        )
+        assert results == [("Me", "Myway 88")]
+        assert len(managed_connector_with_data._unique_results) == 2
+
+        # now reset so fetch starts at the first value again
+        if managed_connector_with_data._driver_is_async:
+            await managed_connector_with_data.reset_async_connections()
+        else:
+            managed_connector_with_data.reset_connections()
+        assert len(managed_connector_with_data._unique_results) == 0
+
+        # ensure it's really reset
+        results = await managed_connector_with_data.fetch_many(
+            "SELECT * FROM customers", size=3
+        )
+        assert results == [
+            ("Marvin", "Highway 42"),
+            ("Ford", "Highway 42"),
+            ("Unknown", "Space"),
+        ]
+        assert len(managed_connector_with_data._unique_results) == 1
+
+    async def test_fetch_all(self, managed_connector_with_data):
+        # test with parameters
+        results = await managed_connector_with_data.fetch_all(
+            "SELECT * FROM customers WHERE address = :address",
+            parameters={"address": "Highway 42"},
+        )
+        expected = [("Marvin", "Highway 42"), ("Ford", "Highway 42")]
+        assert results == expected
+
+        # there should be no more results
+        results = await managed_connector_with_data.fetch_all(
+            "SELECT * FROM customers WHERE address = :address",
+            parameters={"address": "Highway 42"},
+        )
+        assert results == []
+        assert len(managed_connector_with_data._unique_results) == 1
+
+        # now reset so fetch one starts at the first value again
+        if managed_connector_with_data._driver_is_async:
+            await managed_connector_with_data.reset_async_connections()
+        else:
+            managed_connector_with_data.reset_connections()
+        assert len(managed_connector_with_data._unique_results) == 0
+
+        # ensure it's really reset
+        results = await managed_connector_with_data.fetch_all(
+            "SELECT * FROM customers WHERE address = :address",
+            parameters={"address": "Highway 42"},
+        )
+        expected = [("Marvin", "Highway 42"), ("Ford", "Highway 42")]
+        assert results == expected
+        assert len(managed_connector_with_data._unique_results) == 1
+
+    async def test_pickleable(self, managed_connector_with_data):
+        await managed_connector_with_data.fetch_one("SELECT * FROM customers")
+        pkl = cloudpickle.dumps(managed_connector_with_data)
+        assert pkl
+        assert cloudpickle.loads(pkl)
+
+    async def test_close(self, managed_connector_with_data):
+        if managed_connector_with_data._driver_is_async:
+            with pytest.raises(RuntimeError, match="Please use the"):
+                managed_connector_with_data.close()
+        else:
+            managed_connector_with_data.close()  # test calling it twice
+
+    async def test_aclose(self, managed_connector_with_data):
+        if not managed_connector_with_data._driver_is_async:
+            with pytest.raises(RuntimeError, match="Please use the"):
+                await managed_connector_with_data.aclose()
+        else:
+            await managed_connector_with_data.aclose()  # test calling it twice
+
+    async def test_enter(self, managed_connector_with_data):
+        if managed_connector_with_data._driver_is_async:
+            with pytest.raises(RuntimeError, match="cannot be run"):
+                with managed_connector_with_data:
+                    pass
+
+    async def test_aenter(self, managed_connector_with_data):
+        if not managed_connector_with_data._driver_is_async:
+            with pytest.raises(RuntimeError, match="cannot be run"):
+                async with managed_connector_with_data:
+                    pass
